refactor(types): slider
